+++ tests/dataclasses/dataclasses_test.py
@@ -11,9 +11,9 @@ from typing import Optional, Any
 import pytest
 
 from tests.test_utils import UNSET_PARAMETER
+from validataclass.dataclasses import validataclass, validataclass_field, Default, DefaultFactory, DefaultUnset, NoDefault
 from validataclass.exceptions import DataclassValidatorFieldException
-from validataclass.helpers import validataclass, validataclass_field, Default, DefaultFactory, DefaultUnset, NoDefault, \
-    OptionalUnset, UnsetValue
+from validataclass.helpers import OptionalUnset, UnsetValue
 from validataclass.validators import IntegerValidator, StringValidator, Noneable
